@@ -11,51 +11,52 @@
   "dependencies": {
     "@babel/preset-env": "^7.1.6",
     "@expo/vector-icons": "^13.0.0",
-    "@react-native-async-storage/async-storage": "1.17.11",
+    "@react-native-async-storage/async-storage": "1.18.2",
     "@react-native-community/blur": "^4.3.0",
     "@react-native-community/masked-view": "^0.1.11",
-    "@react-native-picker/picker": "2.4.8",
+    "@react-native-picker/picker": "2.4.10",
     "@react-navigation/bottom-tabs": "^6.5.7",
     "@react-navigation/drawer": "^6.6.2",
     "@react-navigation/material-top-tabs": "^6.6.2",
     "@react-navigation/native": "^6.1.6",
     "@react-navigation/stack": "^6.3.16",
     "@rneui/base": "^4.0.0-rc.7",
-    "@types/react": "~18.0.27",
-    "@types/react-native": "~0.70.6",
+    "@types/react": "^18.0.27",
+    "@types/react-native": "^0.70.6",
     "axios": "^1.3.4",
     "base64-js": "^1.5.1",
     "date-fns": "^2.30.0",
-    "expo": "~48.0.9",
-    "expo-constants": "~14.2.1",
-    "expo-image-picker": "~14.1.1",
-    "expo-linear-gradient": "~12.1.2",
-    "expo-status-bar": "~1.4.4",
+    "expo": "^49.0.7",
+    "expo-constants": "~14.4.2",
+    "expo-image-picker": "~14.3.2",
+    "expo-linear-gradient": "~12.3.0",
+    "expo-status-bar": "~1.6.0",
     "native-base": "^3.4.28",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.43.8",
     "react-icons": "^4.8.0",
-    "react-native": "0.71.8",
-    "react-native-gesture-handler": "~2.9.0",
+    "react-native": "0.72.5",
+    "react-native-gesture-handler": "~2.12.0",
     "react-native-image-picker": "^5.6.1",
     "react-native-image-resizer": "^1.4.5",
-    "react-native-pager-view": "6.1.2",
+    "react-native-pager-view": "6.2.0",
     "react-native-paper": "^5.5.2",
-    "react-native-reanimated": "~2.14.4",
+    "react-native-reanimated": "~3.3.0",
     "react-native-responsive-screen": "^1.4.2",
-    "react-native-safe-area-context": "4.5.0",
-    "react-native-screens": "~3.20.0",
-    "react-native-svg": "13.4.0",
+    "react-native-safe-area-context": "4.6.3",
+    "react-native-screens": "~3.22.0",
+    "react-native-svg": "13.9.0",
     "react-native-tab-view": "^3.0.0",
     "react-native-vector-icons": "^9.2.0",
-    "react-native-web": "~0.18.10",
+    "react-native-web": "~0.19.6",
     "react-query": "^3.39.3",
-    "typescript": "^4.6.3"
+    "typescript": "^5.1.3",
+    "uuid": "^9.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.21.3",
-    "@expo/webpack-config": "^18.0.1",
+    "@expo/webpack-config": "^19.0.0",
     "prettier": "^2.8.7"
   },
   "private": true
